refactor wallet and transaction service to reduce repetition,add count and isExist methods to repo,refavtor pagination response for transaction to return count and pages, and untrack readonly queries

+++ ServiceAbstraction/DTOs/ItemToBuyDTOs/SaveMoneyDTO.cs
@@ -11,7 +11,7 @@ namespace ServiceAbstraction.DTOs.ItemToBuyDTOs
     {
         public int Id { get; set; }
         public int walletId {  get; set; }  
-        public int amount { get; set; }
+        public decimal amount { get; set; }
         public MoneySource source {  get; set; }
 
     }
